Fix fhir_table_list which caused many metrics to be skipped in diff test
Because of the capitalization, FHIR resources with multiple words were not found, e.g. CareTeam became Careteam.

Also fix the source for columns in FHIR resource views: don't try to use the view itself while it is being created.

Also crash when table not found in get_datatype_dict to prevent similar errors in the future.

Also remove a too often shown warning.

PiperOrigin-RevId: 595972492
Change-Id: I0c166716b0f875070d432ce50b98cc3277248ae8